Only maximize windows that support it
Previously, all windows were maximized, including dialog windows that
cannot normally be maximized (and cannot be unmaximized by the user).

Additionally, this explicitly unmaximizes any windows that are not
marked to support maximization. This helps to fix windows that remember
their size and position and were previously maximized by this extension.
Such windows might still remember their size and fill the screen, but
this can usually be fixed by resizing the window manually.

This fixes #2