Remove extension version from metadata
AFAIU this version number is intended to be filled with a linear version
number by extensions.gnome.org, so any version in here is not preserved
when uploading to there. Also, for gnome-shell, it seems that the
presence of this version number is an indication that the extension
installation is managed by gnome-shell.

In practice, when using a symlink directly into the git checkout for
this extension, the presence of this version key causes gnome-shell to
(presumably) consider this extension invalid and (observed) replace it
with the latest version downloaded from extensions.gnome.org.